Refactor and add command aliases for find rule commands
