+++ src/main/java/seedu/algobase/logic/parser/findrule/ApplyCommandParser.java
@@ -1,9 +1,11 @@
-package seedu.algobase.logic.parser;
+package seedu.algobase.logic.parser.findrule;
 
 import static seedu.algobase.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.algobase.commons.core.index.Index;
-import seedu.algobase.logic.commands.ApplyCommand;
+import seedu.algobase.logic.commands.findrule.ApplyCommand;
+import seedu.algobase.logic.parser.Parser;
+import seedu.algobase.logic.parser.ParserUtil;
 import seedu.algobase.logic.parser.exceptions.ParseException;
 
 /**
