+++ src/test/java/seedu/algobase/logic/commands/ApplyCommandTest.java
@@ -15,6 +15,7 @@ import org.junit.jupiter.api.Test;
 import seedu.algobase.commons.core.Messages;
 import seedu.algobase.commons.core.index.Index;
 import seedu.algobase.logic.CommandHistory;
+import seedu.algobase.logic.commands.findrule.ApplyCommand;
 import seedu.algobase.model.Model;
 import seedu.algobase.model.ModelManager;
 import seedu.algobase.model.UserPrefs;
