+++ src/test/java/seedu/algobase/logic/commands/DeleteFindRuleCommandTest.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.Test;
 import seedu.algobase.commons.core.Messages;
 import seedu.algobase.commons.core.index.Index;
 import seedu.algobase.logic.CommandHistory;
+import seedu.algobase.logic.commands.findrule.DeleteFindRuleCommand;
 import seedu.algobase.model.Model;
 import seedu.algobase.model.ModelManager;
 import seedu.algobase.model.UserPrefs;
